Fixed OpenGLGraphics::drawRescaledImagePattern
This fixes overlay effects that are meant to scale with screen
resolution.

The problem was that the texture coordinates were not calculated
correctly. They were adjusted to the scaled size of the image, and when
scaling both the vertex and the texture coordinates, the image will
simply not appear scaled at all. Now the texture coordinates are
calculated based on the visible part of the original texture.

There was also a problem with the vertex coordinates, which were not
taking into account the visible part of the image.

TMW-Mantis-issue: 1047
Reviewed-by: Andrei Karas <akaras@inbox.ru>

Conflicts:

	NEWS